Move related code to a separate file
