+++ mimium-lang/src/ast_interpreter.rs
@@ -1,16 +1,13 @@
-use std::{cell::RefCell, collections::BTreeMap};
-
-use id_arena::Arena;
-use string_interner::{backend::StringBackend, StringInterner};
-
 use crate::{
-    ast::{self, Expr, ExprNodeId, NodeId, Symbol, ToNodeId, ToSymbol},
+    ast::{self, Symbol, ToSymbol},
     compiler::{Error as CompileError, ErrorKind},
-    integer, numeric,
+    integer,
+    interner::ExprNodeId,
+    numeric,
     pattern::{TypedId, TypedPattern},
     runtime::builtin_fn,
     string_t,
-    types::{PType, Type, TypeNodeId},
+    types::{PType, Type},
     unit,
     utils::{
         environment::Environment,
@@ -93,73 +90,6 @@ const EXTERN_SYMS: [&str; 28] = [
     "cos", "not", "round", "floor", "ceil", "atan", "sqrt", "abs", "min", "max", "pow", "log",
     "print", "println",
 ];
-
-pub struct SessionGlobals {
-    pub symbol_interner: StringInterner<StringBackend<usize>>,
-    pub expr_storage: Arena<Expr>,
-    pub type_storage: Arena<Type>,
-    pub span_storage: BTreeMap<NodeId, Span>,
-}
-
-impl SessionGlobals {
-    fn store_expr(&mut self, expr: Expr) -> ExprNodeId {
-        ExprNodeId(self.expr_storage.alloc(expr))
-    }
-
-    fn store_type(&mut self, ty: Type) -> TypeNodeId {
-        TypeNodeId(self.type_storage.alloc(ty))
-    }
-
-    fn store_span<T: ToNodeId>(&mut self, node_id: T, span: Span) {
-        self.span_storage.insert(node_id.to_node_id(), span);
-    }
-
-    pub fn store_expr_with_span(&mut self, expr: Expr, span: Span) -> ExprNodeId {
-        let expr_id = self.store_expr(expr);
-        self.store_span(expr_id, span);
-        expr_id
-    }
-
-    pub fn store_type_with_span(&mut self, ty: Type, span: Span) -> TypeNodeId {
-        let type_id = self.store_type(ty);
-        self.store_span(type_id, span);
-        type_id
-    }
-
-    pub fn get_expr(&self, expr_id: ExprNodeId) -> &Expr {
-        self.expr_storage
-            .get(expr_id.0)
-            .expect("Unknown ExprNodeId")
-    }
-
-    pub fn get_type(&self, type_id: TypeNodeId) -> &Type {
-        self.type_storage
-            .get(type_id.0)
-            .expect("Unknown TypeNodeId")
-    }
-
-    pub fn get_span<T: ToNodeId>(&self, node_id: T) -> &Span {
-        self.span_storage
-            .get(&node_id.to_node_id())
-            .expect("Unknown NodeID")
-    }
-}
-
-thread_local!(static SESSION_GLOBALS: RefCell<SessionGlobals> =  RefCell::new(
-    SessionGlobals {
-        symbol_interner: StringInterner::new(),
-        expr_storage: Arena::new(),
-        type_storage: Arena::new(),
-        span_storage: BTreeMap::new()
-    }
-));
-
-pub fn with_session_globals<R, F>(f: F) -> R
-where
-    F: FnOnce(&mut SessionGlobals) -> R,
-{
-    SESSION_GLOBALS.with_borrow_mut(f)
-}
 
 fn eval_literal(e: &ast::Literal) -> Value {
     match e {
