+++ mimium-lang/src/lib.rs
@@ -4,16 +4,18 @@
 #![feature(if_let_guard)]
 #![feature(lazy_cell)]
 // pub mod wcalculus;
-pub(crate) mod pattern;
 pub mod ast;
 pub mod mir;
+pub(crate) mod pattern;
 
 pub mod types;
 pub mod utils;
 
+pub mod audio;
 pub mod compiler;
 pub mod runtime;
-pub mod audio;
+
+pub mod interner;
 
 pub mod ast_interpreter;
 pub mod mir_interpreter;
